tester-service/requirements.txt
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.15
Werkzeug==2.0.1
psycopg2-binary
flask_migrate
kafka-python
flask
flask_cors
aiven-client
python-dotenv
python-dotenv
eventlet
pika
redis
Flask-Executor
flask_coney
schedule